Move gather_elements test to its own module
The gather_elements test was separated from gather/mod.rs and placed in a new module at gather_elements/mod.rs. The test_mod.rs file was updated to include the new module.
