+++ crates/burn-import/onnx-tests/tests/gather/mod.rs
@@ -3,7 +3,6 @@ use crate::include_models;
 include_models!(
     gather_1d_idx,
     gather_2d_idx,
-    gather_elements,
     gather_scalar,
     gather_scalar_out,
     gather_shape,
@@ -87,21 +86,6 @@ mod tests {
 
         let expected = 2.0f32;
         assert_eq!(output, expected);
-    }
-
-    #[test]
-    fn gather_elements() {
-        // Initialize the model with weights (loaded from the exported file)
-        let model: gather_elements::Model<TestBackend> = gather_elements::Model::default();
-
-        let device = Default::default();
-        // Run the model
-        let input = Tensor::<TestBackend, 2>::from_floats([[1., 2.], [3., 4.]], &device);
-        let index = Tensor::<TestBackend, 2, Int>::from_ints([[0, 0], [1, 0]], &device);
-        let output = model.forward(input, index);
-        let expected = TensorData::from([[1f32, 1.], [4., 3.]]);
-
-        assert_eq!(output.to_data(), expected);
     }
 
     #[test]
