+++ crates/burn-import/onnx-tests/tests/test_mod.rs
@@ -45,6 +45,7 @@ pub mod expand;
 pub mod flatten;
 pub mod floor;
 pub mod gather;
+pub mod gather_elements;
 pub mod gelu;
 pub mod gemm;
 pub mod global_avr_pool;
